requirements.txt
# Core ML Libraries
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
accelerate==1.3.0
datasets==3.3.1
diffusers==0.32.2
bitsandbytes==0.42.0

# Vision and Image Processing
pillow==11.1.0
opencv-python-headless==4.8.0.76
qwen-vl-utils==0.0.10

# Scientific Computing
numpy==2.2.2
scipy==1.15.1
scikit-learn==1.6.1
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2

# Jupyter and Development
jupyter==1.1.1
ipython==8.31.0
ipywidgets==8.1.5
notebook==7.3.2
jupyterlab==4.3.4

# Utilities
tqdm==4.67.1
requests==2.32.3
psutil==6.1.1
pyyaml==6.0.2
python-dotenv==1.0.1

# Monitoring and Logging
wandb==0.19.5
neptune==1.13.0

# Cloud and API
openai==1.63.2
boto3==1.36.12

# Audio/Video Processing
av==14.1.0

# Development Tools
pytest==8.3.4
black==24.0.0
flake8==7.0.0
